requirements.txt
futures==2.1.4
psutil==1.0.1
tornado==3.1
ujson==1.33
uuid==1.30
wsgiref==0.1.2
xmpppy==0.5.0rc1
boto==2.33.0
repoze.lru==0.6

